Struct minidump::format::MDRawContextAMD64

pub struct MDRawContextAMD64 {
    pub p1_home: u64,
    pub p2_home: u64,
    pub p3_home: u64,
    pub p4_home: u64,
    pub p5_home: u64,
    pub p6_home: u64,
    pub context_flags: u32,
    pub mx_csr: u32,
    pub cs: u16,
    pub ds: u16,
    pub es: u16,
    pub fs: u16,
    pub gs: u16,
    pub ss: u16,
    pub eflags: u32,
    pub dr0: u64,
    pub dr1: u64,
    pub dr2: u64,
    pub dr3: u64,
    pub dr6: u64,
    pub dr7: u64,
    pub rax: u64,
    pub rcx: u64,
    pub rdx: u64,
    pub rbx: u64,
    pub rsp: u64,
    pub rbp: u64,
    pub rsi: u64,
    pub rdi: u64,
    pub r8: u64,
    pub r9: u64,
    pub r10: u64,
    pub r11: u64,
    pub r12: u64,
    pub r13: u64,
    pub r14: u64,
    pub r15: u64,
    pub rip: u64,
    pub float_save: [u8; 512],
    pub vector_register: [u128; 26],
    pub vector_control: u64,
    pub debug_control: u64,
    pub last_branch_to_rip: u64,
    pub last_branch_from_rip: u64,
    pub last_exception_to_rip: u64,
    pub last_exception_from_rip: u64,
}

Fields
p1_home: u64 p2_home: u64 p3_home: u64 p4_home: u64 p5_home: u64 p6_home: u64 context_flags: u32 mx_csr: u32 cs: u16 ds: u16 es: u16 fs: u16 gs: u16 ss: u16 eflags: u32 dr0: u64 dr1: u64 dr2: u64 dr3: u64 dr6: u64 dr7: u64 rax: u64 rcx: u64 rdx: u64 rbx: u64 rsp: u64 rbp: u64 rsi: u64 rdi: u64 r8: u64 r9: u64 r10: u64 r11: u64 r12: u64 r13: u64 r14: u64 r15: u64 rip: u64 float_save: [u8; 512]

This is defined as a union in the C headers, but also MAXIMUM_SUPPORTED_EXTENSION is defined as 512 bytes.

Callers that want to access the underlying data can use Pread to read either an MDXmmSaveArea32AMD64 or SSERegisters struct from this raw data as appropriate.

vector_register: [u128; 26] vector_control: u64 debug_control: u64 last_branch_to_rip: u64 last_branch_from_rip: u64 last_exception_to_rip: u64 last_exception_from_rip: u64