Struct minidump::Minidump

pub struct Minidump {
    pub header: MDRawHeader,
    // some fields omitted
}

An index into the contents of a minidump.

The Minidump struct represents the parsed header and indices contained at the start of a minidump file. It can be instantiated by calling the Minidump::read or Minidump::read_path methods.

Examples

use minidump::Minidump;
use std::fs::File;

let file = try!(File::open("../testdata/test.dmp"));
let dump = Minidump::read(file).unwrap();

Fields
header: MDRawHeader
Methods

impl Minidump

fn read_path(path: &Path) -> Result<Minidump, Error>

Read a Minidump from a Path to a file on disk.

fn read<T: Readable + 'static>(f: T) -> Result<Minidump, Error>

Read a Minidump from a Readable.

fn get_stream<T: MinidumpStream>(&mut self) -> Result<T, Error>

Get a known stream of data from the minidump.

For streams known to this module whose types implement the MinidumpStream trait, this method allows reading the stream data as a specific type.

fn get_raw_stream(&mut self, stream_type: u32) -> Result<Vec<u8>, Error>

Get a stream of raw data from the minidump.

This can be used to get the contents of arbitrary minidump streams. For streams of known types you almost certainly want to use get_stream instead.

fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a verbose description of the Minidump to f.