Struct minidump::MinidumpBreakpadInfo

pub struct MinidumpBreakpadInfo {
    pub dump_thread_id: Option<u32>,
    pub requesting_thread_id: Option<u32>,
    // some fields omitted
}

Additional information about process state.

MinidumpBreakpadInfo wraps MDRawBreakpadInfo, which is an optional stream in a minidump that provides additional information about the process state at the time the minidump was generated.

Fields
dump_thread_id: Option<u32>

The thread that wrote the minidump.

requesting_thread_id: Option<u32>

The thread that requested that a minidump be written.

Methods

impl MinidumpBreakpadInfo

pub fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpBreakpadInfo to f.

This is very verbose, it is the format used by minidump_dump.