Struct minidump::MinidumpMemoryList

pub struct MinidumpMemoryList { /* fields omitted */ }

A list of memory regions included in a minidump.

Methods

impl MinidumpMemoryList

pub fn new() -> MinidumpMemoryList

Return an empty MinidumpMemoryList.

pub fn from_regions(regions: Vec<MinidumpMemory>) -> MinidumpMemoryList

Create a MinidumpMemoryList from a list of MinidumpMemorys.

pub fn iter<'a>(&'a self) -> MemoryRegions<'a>

Iterate over the memory regions in the order contained in the minidump.

pub fn by_addr<'a>(&'a self) -> MemoryRegions<'a>

Iterate over the memory regions in order by memory address.

pub fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpMemoryList to f.

This is very verbose, it is the format used by minidump_dump.