Struct minidump::MinidumpMiscInfo

pub struct MinidumpMiscInfo {
    pub raw: MDRawMiscInfo,
    pub process_create_time: Option<DateTime<UTC>>,
}

Miscellaneous information about the process that wrote the minidump.

Fields

raw

The MDRawMiscInfo struct direct from the minidump.

process_create_time

When the process started, if available.

Methods

impl MinidumpMiscInfo

fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpMiscInfo to f.

This is very verbose, it is the format used by minidump_dump.