Struct minidump::MinidumpModuleList

pub struct MinidumpModuleList { /* fields omitted */ }

A list of MinidumpModules contained in a Minidump.

Methods

impl MinidumpModuleList

pub fn new() -> MinidumpModuleList

Return an empty MinidumpModuleList.

pub fn from_modules(modules: Vec<MinidumpModule>) -> MinidumpModuleList

Create a MinidumpModuleList from a list of MinidumpModules.

pub fn main_module(&self) -> Option<&MinidumpModule>

Returns the module corresponding to the main executable.

pub fn module_at_address(&self, addr: u64) -> Option<&MinidumpModule>

Return a MinidumpModule whose address range covers addr.

pub fn iter<'a>(&'a self) -> Modules<'a>

Iterate over the modules in arbitrary order.

pub fn by_addr<'a>(&'a self) -> Modules<'a>

Iterate over the modules in order by memory address.

pub fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpModuleList to f.

This is very verbose, it is the format used by minidump_dump.