Struct minidump::ProcessState

pub struct ProcessState {
    pub time: DateTime<UTC>,
    pub process_create_time: Option<DateTime<UTC>>,
    pub crash_reason: Option<CrashReason>,
    pub crash_address: Option<u64>,
    pub assertion: Option<String>,
    pub requesting_thread: Option<usize>,
    pub threads: Vec<CallStack>,
    pub system_info: SystemInfo,
    pub modules: MinidumpModuleList,
}

The state of a process as recorded by a Minidump.

Fields
time: DateTime<UTC>

When the minidump was written.

process_create_time: Option<DateTime<UTC>>

When the process started, if available

crash_reason: Option<CrashReason>

If the process crashed, a CrashReason describing the crash reason.

crash_address: Option<u64>

The memory address implicated in the crash.

If the process crashed, and if the crash reason implicates memory, this is the memory address that caused the crash. For data access errors this will be the data address that caused the fault. For code errors, this will be the address of the instruction that caused the fault.

assertion: Option<String>

A string describing an assertion that was hit, if present.

requesting_thread: Option<usize>

The index of the thread that requested a dump be written. If a dump was produced as a result of a crash, this will point to the thread that crashed. If the dump was produced as by user code without crashing, and the dump contains extended Breakpad information, this will point to the thread that requested the dump. If the dump was not produced as a result of an exception and no extended Breakpad information is present, this field will be None.

threads: Vec<CallStack>

Stacks for each thread (except possibly the exception handler thread) at the time of the crash.

system_info: SystemInfo

Information about the system on which the minidump was written.

modules: MinidumpModuleList

The modules that were loaded into the process represented by the ProcessState.

Methods

impl ProcessState

pub fn crashed(&self) -> bool

true if the minidump was written in response to a process crash.

pub fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of the process state to f.

This is very verbose, it implements the output format used by minidump_stackwalk.