Crate minidump

A parser for the minidump file format.

The minidump module provides a parser for the minidump file format as produced by Microsoft's MinidumpWriteDump API and the Google Breakpad library.

The primary API for this module is the Minidump struct, which can be instantiated by calling the Minidump::read or Minidump::read_path methods.

Modules

format
Minidump structure definitions.

system_info
Information about the system that produced a Minidump.

Structs

MemoryRegions
An iterator over MinidumpMemorys.

Minidump
An index into the contents of a minidump.

MinidumpAssertion
Information about an assertion that caused a crash.

MinidumpBreakpadInfo
Additional information about process state.

MinidumpContext
CPU context such as register states.

MinidumpException
Information about the exception that caused the minidump to be generated.

MinidumpMemory
A region of memory from the process that wrote the minidump.

MinidumpMemoryList
A list of memory regions included in a minidump.

MinidumpMiscInfo
Miscellaneous information about the process that wrote the minidump.

MinidumpModule
An executable or shared library loaded in the process at the time the Minidump was written.

MinidumpModuleList
A list of MinidumpModules contained in a Minidump.

MinidumpSystemInfo
Information about the system that generated the minidump.

MinidumpThread
The state of a thread from the process when the minidump was written.

MinidumpThreadList
A list of MinidumpThreads contained in a Minidump.

Modules
An iterator over MinidumpModules.

Enums

CodeView
CodeView data describes how to locate debug symbols

ContextError
Errors encountered while reading a MinidumpContext.

CrashReason
The reason for a process crash.

Endian
The endianness (byte order) of a stream of bytes

Error
Errors encountered while reading a Minidump.

MinidumpContextValidity
Information about which registers are valid in a MinidumpContext.

MinidumpRawContext
The CPU-specific context structure.

RawMiscInfo

Traits

CPUContext
Generic over the specifics of a CPU context.

MinidumpStream
The fundamental unit of data in a Minidump.

Module
An executable or shared library loaded in a process.

Readable
Shorthand for Read + Seek