Struct minidump::CallStack

pub struct CallStack {
    pub frames: Vec<StackFrame>,
    pub info: CallStackInfo,
}

A stack of StackFrames produced as a result of unwinding a thread.

Fields

frames

The stack frames. By convention, the stack frame at index 0 is the innermost callee frame, and the frame at the highest index in a call stack is the outermost caller.

info

Information about this CallStack.

Methods

impl CallStack

fn with_info(info: CallStackInfo) -> CallStack

Create a CallStack with info and no frames.

fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of the call stack to f.

This is very verbose, it implements the output format used by minidump_stackwalk.