Struct minidump::Minidump

pub struct Minidump<'a, T> where
    T: Deref<Target = [u8]> + 'a,  {
    pub header: MDRawHeader,
    // some fields omitted
}

An index into the contents of a minidump.

The Minidump struct represents the parsed header and indices contained at the start of a minidump file. It can be instantiated by calling the Minidump::read or Minidump::read_path methods.

Examples

use minidump::Minidump;

let dump = Minidump::read_path("../testdata/test.dmp")?;

Fields
header: MDRawHeader
Methods

impl<'a> Minidump<'a, Mmap>

pub fn read_path<P>(path: P) -> Result<Minidump<'a, Mmap>, Error> where
P: AsRef<Path>,

Read a Minidump from a Path to a file on disk.

See the type definition for an example.

impl<'a, T> Minidump<'a, T> where
T: Deref<Target = [u8]> + 'a,

pub fn read(data: T) -> Result<Minidump<'a, T>, Error>

Read a Minidump from the provided data.

Typically this will be a Vec<u8> or &[u8] with the full contents of the minidump, but you can also use something like memmap::Mmap.

pub fn get_stream<'b, S>(&'b self) -> Result<S, Error> where
S: MinidumpStream<'a>,
'b: 'a,

Get a known stream of data from the minidump.

For streams known to this module whose types implement the MinidumpStream trait, this method allows reading the stream data as a specific type.

pub fn get_raw_stream<'b>(&'b self, stream_type: u32) -> Result<&'a [u8], Error> where
'b: 'a,

Get a stream of raw data from the minidump.

This can be used to get the contents of arbitrary minidump streams. For streams of known types you almost certainly want to use get_stream instead.

pub fn print<W: Write>(&self, f: &mut W) -> Result<()>

Write a verbose description of the Minidump to f.