Struct minidump::MinidumpContext

pub struct MinidumpContext {
    pub raw: MinidumpRawContext,
    pub valid: MinidumpContextValidity,
}

CPU context such as register states.

MinidumpContext carries a CPU-specific MDRawContext structure, which contains CPU context such as register states. Each thread has its own context, and the exception record, if present, also has its own context. Note that if the exception record is present, the context it refers to is probably what the user wants to use for the exception thread, instead of that thread's own context. The exception thread's context (as opposed to the exception record's context) will contain context for the exception handler (which performs minidump generation), and not the context that caused the exception (which is probably what the user wants).

Fields
raw: MinidumpRawContext

The raw CPU register state.

valid: MinidumpContextValidity

Which registers are valid in raw.

Methods

impl MinidumpContext

pub fn from_raw(raw: MinidumpRawContext) -> MinidumpContext

Return a MinidumpContext given a MinidumpRawContext.

pub fn read<T: Readable>(
f: &mut T,
location: &MDLocationDescriptor
) -> Result<MinidumpContext, ContextError>

Read a MinidumpContext from a file.

pub fn get_instruction_pointer(&self) -> u64

pub fn get_stack_pointer(&self) -> u64

pub fn format_register(&self, reg: &str) -> String

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn general_purpose_registers(&self) -> &'static [&'static str]

pub fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpContext to f.

This is very verbose, it is the format used by minidump_dump.