Struct minidump::MinidumpException

pub struct MinidumpException {
    pub raw: MDRawExceptionStream,
    pub thread_id: u32,
    pub context: Option<MinidumpContext>,
}

Information about the exception that caused the minidump to be generated.

MinidumpException wraps MDRawExceptionStream, which contains information about the exception that caused the minidump to be generated, if the minidump was generated in an exception handler called as a result of an exception. It also provides access to a MinidumpContext object, which contains the CPU context for the exception thread at the time the exception occurred.

Fields

raw

thread_id

context

Methods

impl MinidumpException

fn get_crash_address(&self, os: OS) -> u64

Get the crash address for an exception.

fn get_crash_reason(&self, os: OS) -> CrashReason

Get the crash reason for an exception.

fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpException to f.

This is very verbose, it is the format used by minidump_dump.