Struct minidump::MinidumpMemory

pub struct MinidumpMemory {
    pub base_address: u64,
    pub size: u64,
    pub bytes: Vec<u8>,
}

A region of memory from the process that wrote the minidump.

Fields

base_address

The starting address of this range of memory.

size

The length of this range of memory.

bytes

The contents of the memory.

Methods

impl MinidumpMemory

fn read<T: Readable>(f: &mut T, desc: &MDMemoryDescriptor) -> Result<MinidumpMemory, Error>

fn get_memory_at_address<T: Copy + Sized>(&self, addr: u64) -> Option<T>

Get mem::size_of::<T>() bytes of memory at addr from this region.

Return None if the requested address range falls out of the bounds of this memory region.

fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpMemory to f.

This is very verbose, it is the format used by minidump_dump.