Struct minidump::MinidumpMemory

pub struct MinidumpMemory<'a> {
    pub desc: MINIDUMP_MEMORY_DESCRIPTOR,
    pub base_address: u64,
    pub size: u64,
    pub bytes: &'a [u8],
}

A region of memory from the process that wrote the minidump.

Fields
desc: MINIDUMP_MEMORY_DESCRIPTOR

The raw MINIDUMP_MEMORY_DESCRIPTOR from the minidump.

base_address: u64

The starting address of this range of memory.

size: u64

The length of this range of memory.

bytes: &'a [u8]

The contents of the memory.

Methods

impl<'a> MinidumpMemory<'a>

pub fn read(
desc: &MINIDUMP_MEMORY_DESCRIPTOR,
data: &'a [u8]
) -> Result<MinidumpMemory<'a>, Error>

pub fn get_memory_at_address<T>(&self, addr: u64) -> Option<T> where
T: TryFromCtx<'a, Endian, [u8], Error = Error, Size = usize>,
T: SizeWith<Endian, Units = usize>,

Get mem::size_of::<T>() bytes of memory at addr from this region.

Return None if the requested address range falls out of the bounds of this memory region.

pub fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpMemory to f.

This is very verbose, it is the format used by minidump_dump.

pub fn print_contents<T: Write>(&self, f: &mut T) -> Result<()>

Write the contents of this MinidumpMemory to f as a hex string.