Struct minidump::MinidumpModule

pub struct MinidumpModule {
    pub raw: MDRawModule,
    pub codeview_info: Option<CodeView>,
    pub misc_info: Option<MDImageDebugMisc>,
    // some fields omitted
}

An executable or shared library loaded in the process at the time the Minidump was written.

Fields

raw

The MDRawModule direct from the minidump file.

codeview_info

A CodeView record, if one is present.

misc_info

A misc debug record, if one is present.

Methods

impl MinidumpModule

fn new(base: u64, size: u32, name: &str) -> MinidumpModule

Create a MinidumpModule with some basic info.

Useful for testing.

fn read<T: Readable>(f: &mut T, raw: MDRawModule) -> Result<MinidumpModule, Error>

fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpModule to f.

This is very verbose, it is the format used by minidump_dump.