Struct minidump::MinidumpModuleList

pub struct MinidumpModuleList {
    pub modules: Vec<MinidumpModule>,
    // some fields omitted
}

A list of MinidumpModules contained in a Minidump.

Fields

modules

The modules, in the order they were stored in the minidump.

Methods

impl MinidumpModuleList

fn new() -> MinidumpModuleList

Return an empty MinidumpModuleList.

fn from_modules(modules: Vec<MinidumpModule>) -> MinidumpModuleList

Create a MinidumpModuleList from a list of MinidumpModules.

fn main_module(&self) -> Option<&MinidumpModule>

Returns the module corresponding to the main executable.

fn module_at_address(&self, addr: u64) -> Option<&MinidumpModule>

Return a MinidumpModule whose address range covers addr.

fn by_addr(&self) -> Vec<&MinidumpModule>

Return modules in order by memory address.

fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpModuleList to f.

This is very verbose, it is the format used by minidump_dump.