Struct minidump::MinidumpThreadList

pub struct MinidumpThreadList<'a> {
    pub threads: Vec<MinidumpThread<'a>>,
    // some fields omitted
}

A list of MinidumpThreads contained in a Minidump.

Fields
threads: Vec<MinidumpThread<'a>>

The threads, in the order they were present in the Minidump.

Methods

impl<'a> MinidumpThreadList<'a>

pub fn get_thread(&self, id: u32) -> Option<&MinidumpThread<'a>>

Get the thread with id id from this thread list if it exists.

pub fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpThreadList to f.

This is very verbose, it is the format used by minidump_dump.