Struct minidump::MinidumpThreadList

pub struct MinidumpThreadList {
    pub threads: Vec<MinidumpThread>,
    // some fields omitted
}

A list of MinidumpThreads contained in a Minidump.

Fields
threads: Vec<MinidumpThread>

The threads, in the order they were present in the Minidump.

Methods

impl MinidumpThreadList

fn get_thread(&self, id: u32) -> Option<&MinidumpThread>

Get the thread with id id from this thread list if it exists.

fn print<T: Write>(&self, f: &mut T) -> Result<()>

Write a human-readable description of this MinidumpThreadList to f.

This is very verbose, it is the format used by minidump_dump.