Trait minidump::CPUContext

pub trait CPUContext {
    type Register: LowerHex;
    fn get_register_always(&self, reg: &str) -> Self::Register;

    fn get_register(
        &self, 
        reg: &str, 
        valid: &MinidumpContextValidity
    ) -> Option<Self::Register> { ... }

    fn format_register(&self, reg: &str) -> String { ... }
}

Generic over the specifics of a CPU context.

Associated Types

type Register: LowerHex

The word size of general-purpose registers in the context.

Required Methods

fn get_register_always(&self, reg: &str) -> Self::Register

Get a register value regardless of whether it is valid.

Provided Methods

fn get_register(
&self,
reg: &str,
valid: &MinidumpContextValidity
) -> Option<Self::Register>

Get a register value if it is valid.

Get the value of the register named reg from this CPU context if valid indicates that it has a valid value, otherwise return None.

fn format_register(&self, reg: &str) -> String

Return a String containing the value of reg formatted to its natural width.