Trait minidump::MinidumpStream

pub trait MinidumpStream<'a>: Sized {
    const STREAM_TYPE: MINIDUMP_STREAM_TYPE;

    fn read(
        bytes: &'a [u8], 
        all: &'a [u8], 
        endian: Endian
    ) -> Result<Self, Error>;
}

The fundamental unit of data in a Minidump.

Associated Constants

const STREAM_TYPE: MINIDUMP_STREAM_TYPE

The stream type constant used in the md::MDRawDirectory entry.

Required Methods

fn read(bytes: &'a [u8], all: &'a [u8], endian: Endian) -> Result<Self, Error>

Read this MinidumpStream type from bytes.

bytes is the contents of this specific stream. all refers to the full contents of the minidump, for reading auxilliary data referred to with MINIDUMP_LOCATION_DESCRIPTORs.